requirements.txt
Flask==0.12.3
RPi.GPIO==0.7.0
schedule==0.6.0